--- progex3/pubsub/src/controllers/ClientControllerInterface.java
@@ -1,5 +1,0 @@
-package controllers;
-
-public interface ClientControllerInterface {
-	public void onMessageSend(String message, String[] tags);
-}
